refactor sqlite query errors handling
